add deployment environmen

+++ app.mjs
@@ -20,7 +20,8 @@ const today = new Date() ;
 const options = { weekday:'long' , year : 'numeric' , month : 'long' , day : "numeric"} ; 
 const date = today.toLocaleDateString("en-US",options) ; 
 
-connect('mongodb://127.0.0.1:27017/todolistdb')
+//connect('mongodb://127.0.0.1:27017/todolistdb')
+connect('mongodb+srv://atiqur88:test123@cluster0.n6wsmz4.mongodb.net/todolistdb')
 .then(() => {
     console.log('Connected to MongoDB');
   }); // connecting to database todolistdb
@@ -57,7 +58,7 @@ const list = new model('list' , listSchema) ;
 
 
 app.get('/:reqParam', function(req,res){
-    
+    console.log('userdefined route');
     list.findOne({name : req.params.reqParam})
     .then(result => {
         
@@ -78,13 +79,14 @@ app.get('/:reqParam', function(req,res){
 
 
 app.get('/' ,  function(req,res){
+    console.log('home route');
     item.find()
     .then ( result => {
         if (result.length === 0 ){
             item.insertMany(defualtItems).catch(function(error){
                 console.log("error occured" + error) ; 
             });
-            
+            res.redirect('/');
         }
         else {
             res.render('index' , {item_list_name : result , title : date }) ; 
@@ -99,28 +101,61 @@ app.get('/' ,  function(req,res){
 
  
 app.post('/' , function(req , res){
+        const title = req.body.button ; 
+        if (title === date) {
+            const newitem  = new item({
+                name : req.body.list_item 
+            });
+            newitem.save(); 
+            res.redirect('/');
 
-        const newitem  = new item({
-            name : req.body.list_item 
-        });
-        newitem.save(); 
+        }
+        else{
+            list.findOne({name : title})
+            .then ( foundlist => {
+                foundlist.updateOne({$push : {listName : {name : req.body.list_item}}})
+                .then(res => {
+                    console.log('success');
+                });
+                res.redirect('/'+title);
+            })
+            .catch(error => {
+                console.log("not inserted") ; 
+            })
+        }
+        
     
      
-    if ( req.body.name === 'Work list'){ 
-        res.redirect("/work"); 
-    }
-    else {
-        res.redirect("/");
-    }
+    // if ( req.body.name === 'Work list'){ 
+    //     res.redirect("/work"); 
+    // }
+    // else {
+    //     res.redirect("/");
+    // }
     
 });
 app.post("/delete" , function (req, res) {
-    item.deleteOne({name : req.body.checked }).catch(error => {
-        console.log(error); 
-    })
-    res.redirect("/");
+    const listname = req.body.listName; 
+    console.log(listname);
+    if (listname === date){
+        item.deleteOne({name : req.body.checked }).catch(error => {
+            console.log(error); 
+        })
+        res.redirect("/");
+    }
+    else {
+        list.updateOne( {name : listname} ,{$pull : { listName : {name : req.body.checked }}})
+        .then (res => {
+            console.log('success' + res );
+        })
+        .catch(error => {
+            console.log('fail');
+        }) 
+        res.redirect('/'+listname);
+    }
+    
 })
-
-app.listen(process.env.port || 3000 , function(){
-    console.log('app is listening on port 3000');
+const port = process.env.PORT || 3000 ; 
+app.listen(port , function(){
+    console.log('app is listening on port ' + port );
 });
